Move additional classes out of the training file and rename it to trainer.py.
